Move the modules that directly relate to a suite
