+++ packages/vest/src/core/suite/register/spec.js
